Refactored retargeting and implemented AutoRebase2QuditGatePass
