+++ bqskit/passes/__init__.py
@@ -42,9 +42,17 @@ synthesis pass to convert the circuit to native gates.
 
     ExhaustiveGateRemovalPass
     IterativeScanningGateRemovalPass
-    Rebase2QuditGatePass
     ScanningGateRemovalPass
     SubstitutePass
+
+.. rubric:: Retargeting Passes
+
+.. autosummary::
+    :toctree: autogen
+    :recursive:
+
+    AutoRebase2QuditGatePass
+    Rebase2QuditGatePass
 
 .. rubric:: Control Passes
 
@@ -215,9 +223,10 @@ from bqskit.passes.partitioning.scan import ScanPartitioner
 from bqskit.passes.partitioning.single import GroupSingleQuditGatePass
 from bqskit.passes.processing.exhaustive import ExhaustiveGateRemovalPass
 from bqskit.passes.processing.iterative import IterativeScanningGateRemovalPass
-from bqskit.passes.processing.rebase import Rebase2QuditGatePass
 from bqskit.passes.processing.scan import ScanningGateRemovalPass
 from bqskit.passes.processing.substitute import SubstitutePass
+from bqskit.passes.retarget.auto import AutoRebase2QuditGatePass
+from bqskit.passes.retarget.two import Rebase2QuditGatePass
 from bqskit.passes.rules.ch2cnot import CHToCNOTPass
 from bqskit.passes.rules.cnot2ch import CNOTToCHPass
 from bqskit.passes.rules.cnot2cy import CNOTToCYPass
@@ -345,4 +354,5 @@ __all__ = [
     'SubtopologySelectionPass',
     'PermutationAwareSynthesisPass',
     'ToVariablePass',
+    'AutoRebase2QuditGatePass',
 ]
