+++ bqskit/passes/processing/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 
 from bqskit.passes.processing.exhaustive import ExhaustiveGateRemovalPass
 from bqskit.passes.processing.iterative import IterativeScanningGateRemovalPass
-from bqskit.passes.processing.rebase import Rebase2QuditGatePass
+from bqskit.passes.processing.rebase import Rebase2QuditGatePass  # TODO: Remove
 from bqskit.passes.processing.scan import ScanningGateRemovalPass
 from bqskit.passes.processing.substitute import SubstitutePass
